refactor(eval-runtime): 将召回弃答示例收敛为单文件
