+++ docs/reference/eval-runtime-api.md
@@ -168,7 +168,7 @@ Use summary `statistic: 'rate'` for these metrics. The opposite population is `m
 
 The versioned `omk.eval-runtime.abstention/v1` instrument seals pointers, metric IDs and the successful-empty-list protocol into Definition and Runtime identity. It does not change retrieval v1 formulas, scoring, prompts or Core schemas. Do not compare this metric directly with historical scores that count empty Gold as perfect recall; establish an explicit baseline for the new instrument. Natural-language refusals and safety-filter responses are outside this list protocol.
 
-The complete offline example is `examples/eval-runtime/retrieval-abstention.mjs`, with host-owned helpers in `retrieval-abstention-support.mjs`. After building the repository, run `node examples/eval-runtime/retrieval-abstention.mjs`. A separate service can copy both example files and install OMK; no credentials or business network are needed for the synthetic Executor.
+The complete offline example is the single file `examples/eval-runtime/retrieval-abstention.mjs`, including its host-owned helper functions. After building the repository, run `node examples/eval-runtime/retrieval-abstention.mjs`. A separate service can copy this example file and install OMK; no credentials or business network are needed for the synthetic Executor.
 
 The example prepares and validates labels before execution, defaults to rejecting pending samples, and permits explicit exclusion with an audit of source revision, counts and reasons. Business `quality.reviewStatus` is interpreted only by that helper. The actual selected Dataset is sealed by Core. It combines built-in retrieval and abstention with an independent Custom Evaluator for forbidden-ID hits. All samples carry a relevance binding. An empty relevant set on negative samples retains the existing retrieval v1 `invalid` observation; retrieval analyses explicitly select the `answerable` cohort and exclude those negatives before aggregation. Abstention remains applicable to its own populations. The independent forbidden-hit example checks the first three final results, includes only valid samples with a nonempty forbidden list, and makes no claim that this list enumerates all wrong results. A correct retrieval and a forbidden hit can coexist.
